ml-models/requirements.txt
tensorflow==2.13.0
torch==2.1.0
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
mlflow==2.8.1
optuna==3.4.0
prophet==1.1.5
statsmodels==0.14.0
geopy==2.4.0
shapely==2.0.2
folium==0.15.0
psycopg2-binary==2.9.9
scipy==1.10.1
networkx==3.1
